+++ src/teehr/models/loading/utils.py
@@ -7,12 +7,16 @@ class ChunkByEnum(str, Enum):
     """ChunkByEnum."""
     day = "day"
     location_id = "location_id"
+    week = "week"
+    month = "month"
+    year = "year"
 
 
 class SupportedNWMRetroVersionsEnum(str, Enum):
     """SupportedNWMRetroVersionsEnum."""
     nwm20 = "nwm20"
     nwm21 = "nwm21"
+    nwm30 = "nwm30"
 
 
 class SupportedNWMOperationalVersionsEnum(str, Enum):
@@ -34,6 +38,13 @@ class SupportedKerchunkMethod(str, Enum):
     remote = "remote"
     auto = "auto"
 
+
+class SupportedNWMRetroDomainsEnum(str, Enum):
+    """SupportedNWMRetroDomainsEnum."""
+    CONUS = "CONUS"
+    Alaska = "Alaska"
+    PR = "PR"
+    Hawaii = "Hawaii"
 
 # class NWM20RetroConfig(BaseModel):
 #     URL = 's3://noaa-nwm-retro-v2-zarr-pds'