+++ tests/test_teehr_dataset_db.py
@@ -2,6 +2,7 @@
 from pathlib import Path
 import numpy as np
 import pandas as pd
+import geopandas as gpd
 
 from teehr.database.teehr_dataset import TEEHRDatasetDB
 
@@ -323,6 +324,35 @@ def test_get_joined_timeseries_schema():
     pass
 
 
+def test_joined_timeseries_query_gdf():
+    """Test the get joined timeseries as gdf query."""
+    if DATABASE_FILEPATH.is_file():
+        DATABASE_FILEPATH.unlink()
+
+    tds = TEEHRDatasetDB(DATABASE_FILEPATH)
+
+    # Perform the join and insert into duckdb database
+    tds.insert_joined_timeseries(
+        primary_filepath=PRIMARY_FILEPATH,
+        secondary_filepath=SECONDARY_FILEPATH,
+        crosswalk_filepath=CROSSWALK_FILEPATH,
+        drop_added_fields=True,
+    )
+
+    tds.insert_geometry(GEOMETRY_FILEPATH)
+
+    order_by = ["primary_location_id"]
+
+    df = tds.get_joined_timeseries(
+        order_by=order_by,
+        return_query=False,
+        include_geometry=True,
+    )
+
+    assert df.index.size == 216
+    assert isinstance(df, gpd.GeoDataFrame)
+
+
 def test_timeseries_query():
     """Test the get timeseries query."""
     if DATABASE_FILEPATH.is_file():
@@ -337,12 +367,20 @@ def test_timeseries_query():
         crosswalk_filepath=CROSSWALK_FILEPATH,
         drop_added_fields=True,
     )
+    filters = [
+        {
+            "column": "configuration",
+            "operator": "=",
+            "value": "test_short",
+        },
+    ]
 
     order_by = ["primary_location_id"]
 
     df = tds.get_timeseries(
         order_by=order_by,
         timeseries_name="primary",
+        filters=filters,
         return_query=False
     )
 
@@ -364,7 +402,13 @@ def test_timeseries_char_query():
         drop_added_fields=True,
     )
 
-    filters = []
+    filters = [
+        {
+            "column": "configuration",
+            "operator": "=",
+            "value": "test_short",
+        },
+    ]
     group_by = ["primary_location_id"]
     order_by = ["primary_location_id"]
     timeseries_name = "primary"  # "primary, secondary"
@@ -383,12 +427,13 @@ def test_timeseries_char_query():
 if __name__ == "__main__":
 
     test_insert_joined_timeseries()
-    test_join_attributes()
-    test_unique_field_values()
-    test_describe_inputs()
-    test_calculate_field()
     test_unique_field_values()
     test_metrics_query()
+    test_metrics_query_config_filter()
+    test_describe_inputs()
+    test_calculate_field()
+    test_join_attributes()
     test_get_joined_timeseries_schema()
+    test_joined_timeseries_query_gdf()
     test_timeseries_query()
     test_timeseries_char_query()